Доделаны 2 эндпоинта, подключена БД

+++ repository/db_helper.py
@@ -1,4 +1,3 @@
-import asyncio
 from typing import AsyncGenerator
 from sqlalchemy.ext.asyncio import (
     create_async_engine,
@@ -25,9 +24,11 @@ class DataBaseHelper:
             expire_on_commit=False,
         )
 
-    async def init_db(self) -> None:
+    async def init_db(self, is_drop: bool = False) -> None:
         async with self.engine.begin() as conn:
-            await conn.run_sync(Base.metadata.drop_all)
+            if is_drop:
+                await conn.run_sync(Base.metadata.drop_all)
+
             await conn.run_sync(Base.metadata.create_all)
 
     async def get_session(self) -> AsyncGenerator[AsyncSession, Exception]:
@@ -36,19 +37,3 @@ class DataBaseHelper:
                 yield session
         except Exception as e:
             raise e
-
-
-db_helper = DataBaseHelper(engine_url="sqlite+aiosqlite:///database.db")
-asyncio.run(db_helper.init_db())
-# async def create_user():
-#     async with db.session_factory() as session:
-#         new_user = WalletInfo(
-#             address="sdfsfsfdfs", balance=0.0, bandwidth=121241, energy=12414
-#         )
-#         session.add(new_user)
-#         await session.commit()
-#         return new_user
-#
-#
-# asyncio.run(db.init_db())
-# # asyncio.run(create_user())
